Update seed to new jsonforms root component
This updates the seed to use the new root component.
- Removed the store
- move necessary parameters into app.component
- update to jsonform 2.5.0-alpha.0
