+++ package.json
@@ -23,9 +23,9 @@
     "@angular/platform-browser": "^9.0.2",
     "@angular/platform-browser-dynamic": "^9.0.2",
     "@angular/router": "^9.0.2",
-    "@jsonforms/angular": "^2.4.1",
-    "@jsonforms/angular-material": "^2.4.1",
-    "@jsonforms/core": "^2.4.1",
+    "@jsonforms/angular": "^2.5.0-alpha.0",
+    "@jsonforms/angular-material": "^2.5.0-alpha.0",
+    "@jsonforms/core": "^2.5.0-alpha.0",
     "core-js": "^2.4.1",
     "json-refs": "^3.0.4",
     "libphonenumber-js": "^1.6.8",
